requirements.txt
openai
python-dotenv
pandas
tqdm
pyyaml
pytest
